translator: refactor image processing
Adjusting the image processing logic to perform common tasks in the base
translator instead of everything in the storage format translator. This
is in preparation to support new prospect translator types.
